+++ src/main/java/kamitesque/root/Main.java
@@ -7,6 +7,7 @@ import kamitesque.network.packets.KTNetwork;
 import kamitesque.network.proxy.CommonProxy;
 import kamitesque.util.HoeCache;
 import net.minecraftforge.fml.common.Loader;
+import net.minecraftforge.fml.common.MissingModsException;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.SidedProxy;
@@ -17,10 +18,11 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 @Mod(modid = Main.MODID, dependencies =
         "required-after:thaumcraft;" +
         "required-after:thaumicaugmentation;" +
-        "after:kami;"
+        "required-after:thaumicapi;" +
+        "after:kami;" +
 //        "after:new-crimson-revelations;" +
 //        "after:isorropia;" +
-//        "after:thaumictinkerer;" +
+        "after:thaumictinkerer;"
 //        "after:planarartifice;" +
 //        "after:thaumicbases;" +
 //        "after:thaumicforever;" +
@@ -31,7 +33,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 public class Main {
     public static final String MODID = "kamitesque";
     public static final String NAME = "Kamitesque";
-    public static final String VERSION = "0.1.3-BETA";
+    public static final String VERSION = "0.1.4";
 
     @SidedProxy(clientSide = "kamitesque.network.proxy.ClientProxy", serverSide = "kamitesque.network.proxy.CommonProxy")
     public static CommonProxy proxy;
@@ -42,9 +44,9 @@ public class Main {
     @EventHandler
     public void preInit(FMLPreInitializationEvent event) {
 
-        checkDependencies();
+        checkDependencies(event);
 
-        KTNetwork.preInitPackets();
+        KTNetwork.preInitPacketsServer();
         KTTiles.preInitTiles();
         KTEntities.preInitEntities();
 
@@ -67,10 +69,18 @@ public class Main {
 
     }
 
-    private void checkDependencies() {
-//        if (!Loader.isModLoaded("kami") && !Loader.isModLoaded("thaumictinkerer")
-//        ) {
-//            throw new MissingKamiAddonException();
-//        }
+    private void checkDependencies(FMLPreInitializationEvent event) {
+
+        if (!isRebornLoaded() && !isUnofficialLoaded()) {
+            throw new MissingModsException("Kami: Reborn / Thaumic Tinkerer Unofficial", "kami-1.0.6 or above / thaumictinkerer-1.12.2-5.9.15 or above");
+        }
+    }
+
+    public static boolean isRebornLoaded() {
+        return Loader.isModLoaded("kami");
+    }
+
+    public static boolean isUnofficialLoaded() {
+        return Loader.isModLoaded("thaumictinkerer");
     }
 }